Install AWS Forwarder
As per https://docs.datadoghq.com/serverless/forwarder/ the Datadog
Forwarder is an AWS Lambda to ship logs, custom metrics and traces to a
Datadog org. See page for more info.

Bear in mind this just sets up the Lambda and does not configure any
forwarding, that is done outside of this module at this time.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>